+++ shared/src/commonMain/kotlin/ui/navigation/routes/Route.kt
@@ -4,3 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 open class Route()
+
+@Serializable
+object InitializationRoute: Route()