+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Release 5.7.4 (unreleased):
  * Presentation: Always show requested attributes
+ * Add CapabilitiesService, CapabilitiesView to hint wrong settings
+ * Add InitializationView
 
 # Release 5.7.3
  * Update to VC-K 5.8.0, fixing optional attributes during presentation